Fix where_should_be_participating
